+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Polyverse Boost ReST API (Backend)
 
 # Release Notes
 
-## Version 0.12.5: March 5th, 2024
+## Version 0.12.5: March 7th, 2024
 
 ### New Features
 - N/A
@@ -14,6 +14,7 @@ Polyverse Boost ReST API (Backend)
 ### Bug Fixes
 - Fix guidelines being lost during Project creation
 - Fix Project Status reporting synchronized when Generators are in Complete stage but Error'ed
+- Fix local dispatch issue with API Gateway RESPONSE_STREAM support - where HTTP response is embedded in a JSON object
 
 ## Version 0.12.4: March 5th, 2024